refactor: seperate components and use typescript
