+++ package.json
@@ -34,6 +34,11 @@
     "devDependencies": {
         "@types/react": "^17.0.38",
         "@types/react-dom": "^17.0.11",
+        "@types/wordpress__blocks": "^11.0.5",
+        "@types/wordpress__components": "^19.10.1",
+        "@types/wordpress__compose": "^4.0.1",
+        "@types/wordpress__core-data": "^2.4.5",
+        "@types/wordpress__data": "^6.0.1",
         "@wordpress/env": "^4",
         "@wordpress/scripts": "^23.1.0",
         "autoprefixer": "^9.5.0",
