--- src/NextGen/DonationForm/Blocks/DonationFormBlock/block.js
@@ -1,10 +1,0 @@
-import schema from './block.json';
-import {registerBlockType} from '@wordpress/blocks';
-import Edit from './editor/edit';
-
-/**
- * @since 1.0.0
- */
-registerBlockType(schema, {
-    edit: Edit,
-});
