+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/editor/components/Logo.jsx
@@ -1,0 +1,7 @@
+import logo from '../images/givewp-logo.svg';
+
+export default function Logo() {
+    return <div className="givewp-form-block__logo--container">
+        <img className="givewp-form-block__logo" src={logo} alt="givewp-logo"/>
+    </div>
+}
